LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/mime-sniffing.https-expected.txt


FAIL The response from service worker should be correctly MIME siniffed. assert_equals: expected "test" but got ""
